Added more coverage to ethereum.battlefield
- Fixed compilation of contracts and updated minimum required Solidity version.
- Added tests for callcode and improve call covering by testing failure of each call
 - Still misses revertion of each call, but might now be quite possible in all cases.
- Added tests for precompiled contracts.

+++ contract/src/UniswapV2Migrator.sol
@@ -1,12 +1,12 @@
 pragma solidity =0.6.6;
 
-import '../libraries/TransferHelper.sol';
+import "./libraries/TransferHelper.sol";
 
-import './interfaces/IUniswapV2Migrator.sol';
-import './interfaces/V1/IUniswapV1Factory.sol';
-import './interfaces/V1/IUniswapV1Exchange.sol';
-import './interfaces/IUniswapV2Router01.sol';
-import './interfaces/IERC20.sol';
+import "./interfaces/IUniswapV2Migrator.sol";
+import "./interfaces/V1/IUniswapV1Factory.sol";
+import "./interfaces/V1/IUniswapV1Exchange.sol";
+import "./interfaces/IUniswapV2Router01.sol";
+import "./interfaces/IERC20.sol";
 
 contract UniswapV2Migrator is IUniswapV2Migrator {
     IUniswapV1Factory immutable factoryV1;
@@ -21,16 +21,27 @@ contract UniswapV2Migrator is IUniswapV2Migrator {
     // but it's not possible because it requires a call to the v1 factory, which takes too much gas
     receive() external payable {}
 
-    function migrate(address token, uint amountTokenMin, uint amountETHMin, address to, uint deadline)
-        external
-        override
-    {
+    function migrate(
+        address token,
+        uint256 amountTokenMin,
+        uint256 amountETHMin,
+        address to,
+        uint256 deadline
+    ) external override {
         IUniswapV1Exchange exchangeV1 = IUniswapV1Exchange(factoryV1.getExchange(token));
-        uint liquidityV1 = exchangeV1.balanceOf(msg.sender);
-        require(exchangeV1.transferFrom(msg.sender, address(this), liquidityV1), 'TRANSFER_FROM_FAILED');
-        (uint amountETHV1, uint amountTokenV1) = exchangeV1.removeLiquidity(liquidityV1, 1, 1, uint(-1));
+        uint256 liquidityV1 = exchangeV1.balanceOf(msg.sender);
+        require(
+            exchangeV1.transferFrom(msg.sender, address(this), liquidityV1),
+            "TRANSFER_FROM_FAILED"
+        );
+        (uint256 amountETHV1, uint256 amountTokenV1) = exchangeV1.removeLiquidity(
+            liquidityV1,
+            1,
+            1,
+            uint256(-1)
+        );
         TransferHelper.safeApprove(token, address(router), amountTokenV1);
-        (uint amountTokenV2, uint amountETHV2,) = router.addLiquidityETH{value: amountETHV1}(
+        (uint256 amountTokenV2, uint256 amountETHV2, ) = router.addLiquidityETH{value: amountETHV1}(
             token,
             amountTokenV1,
             amountTokenMin,
